add support for private git repos and npm registries
Private git repos can be accessed by providing path to .netrc file. The
file needs to be added to extra-sandbox-paths, and if daemon is used
appropriate permissions need to be set on it.

Private npm registries are supported via tokens passed in as files
(which need to be added to sandbox as above), which are then
turned into env variables. The env variable can then be used in
.yarnrc.yml to set npmAuthToken via env variable interpolation