separate deps from main project
When used with pnp, we can avoid copying deps to the
main project, so that rebuilds are faster and less
spaces is used when only project source changes
(but not deps).

To make this work, we need to patch .pnp.cjs, since
unfortunately symlinks don't work as packageLocations cannot be (see: https://github.com/yarnpkg/berry/issues/3514). We also cannot use
absolute paths, so we end up using relative paths to store locations
containing each deps. This also works for unplugged deps.